requirements.txt
flask>=2.0.1,<2.1.0
Werkzeug>=2.2.3,<2.3.0
opencv-python>=4.5.5.64
pillow>=9.5.0
numpy>=1.26.0
python-magic>=0.4.27
python-magic-bin>=0.4.14; platform_system=='Windows'
exifread>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
scikit-image>=0.20.0
scipy>=1.10.1
pywavelets>=1.4.1
tensorflow-cpu>=2.15.0
dlib>=19.24.0; platform_system!='Windows'
face-recognition>=1.3.0; platform_system!='Windows'
imagehash>=4.3.1
scikit-learn>=1.2.2
matplotlib>=3.7.1
pandas>=2.0.1
tqdm>=4.65.0
PyYAML>=6.0
Flask-WTF>=1.1.1
Flask-RESTful>=0.3.10
gunicorn>=20.1.0
pytest>=7.3.1
pytest-cov>=4.1.0
python-jose>=3.3.0
Faker>=18.7.0
nltk>=3.8.1
transformers>=4.29.2
beautifulsoup4>=4.12.2
lxml>=4.9.2
pefile>=2023.2.7
joblib>=1.2.0 